Adds "Open in Finder" button in Data Storage Folder
Allows users to easily open the data storage folder in their system's file explorer.

This change introduces a button that, when clicked, opens the directory specified in the data path input field, creating the directory if it doesn't already exist.
This improves user experience by providing quick access to the data storage location.